# Improve `cupy.in1d`