New implementation is based on numpy code. This is much much faster and less less memory hungry. Call to `broadcast_to` is avoided now. Additionally `assume_unique` keyword is handled.