@leofang 2020-09-23
Very interesting! Could you share insights on why the NumPy approach works better (which is not obvious to me), and also post performance benchmarks that you did before and with this PR?