@mrkwjc 2020-09-23
Numpy uses approach with sorting input arrays, whereas cupy tries direct comparison by broadcasting inputs. Broadcasting is generally ok, it creates strided views on inputs, but the next comparison line tries to allocate MxN boolean array, where M, N are lengths of inputs. This is a recipe for out of memory error:

```python
In [11]: import cupy

In [12]: a = cupy.random.randint(0, 1000000, size=500000)

In [13]: b = cupy.random.randint(0, 1000000, size=100000)

In [14]: cupy.in1d(a, b)
---------------------------------------------------------------------------
OutOfMemoryError                          Traceback (most recent call last)
<ipython-input-14-e28ce4babb52> in <module>
----> 1 cupy.in1d(a, b)

~/miniconda3/lib/python3.8/site-packages/cupy/logic/truth.py in in1d(ar1, ar2, assume_unique, invert)
     92     ar1_broadcast = cupy.broadcast_to(ar1[..., cupy.newaxis], shape)
     93     ar2_broadcast = cupy.broadcast_to(ar2, shape)
---> 94     count = (ar1_broadcast == ar2_broadcast).sum(axis=1)
     95     if invert:
     96         return count == 0

cupy/core/core.pyx in cupy.core.core.ndarray.__richcmp__()

cupy/core/_kernel.pyx in cupy.core._kernel.ufunc.__call__()

cupy/core/_kernel.pyx in cupy.core._kernel._get_out_args()

cupy/core/core.pyx in cupy.core.core.ndarray.__init__()

cupy/cuda/memory.pyx in cupy.cuda.memory.alloc()

cupy/cuda/memory.pyx in cupy.cuda.memory.MemoryPool.malloc()

cupy/cuda/memory.pyx in cupy.cuda.memory.MemoryPool.malloc()

cupy/cuda/memory.pyx in cupy.cuda.memory.SingleDeviceMemoryPool.malloc()

cupy/cuda/memory.pyx in cupy.cuda.memory.SingleDeviceMemoryPool._malloc()

cupy/cuda/memory.pyx in cupy.cuda.memory._try_malloc()

OutOfMemoryError: Out of memory allocating 50,000,000,000 bytes (allocated so far: 18,521,088 bytes).
```
Nothing like this happens with numpy's implementation, also for many times bigger inputs. In terms of speed current in1d is poor for larger inputs:
```python
In [15]: a = cupy.random.randint(0, 100000, size=50000)

In [16]: b = cupy.random.randint(0, 100000, size=50000)

In [17]: %timeit cupy.in1d(a, b)
45.7 ms ± 37.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [18]: %timeit in1d(a, b)  # Numpy-like implementation
1.78 ms ± 72.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
For smaller arrays `cupy.in1d` is faster:
```python
In [19]: b = cupy.random.randint(0, 100000, size=5000)

In [20]: a = cupy.random.randint(0, 100000, size=5000)

In [21]: timeit cupy.in1d(a, b)
490 µs ± 112 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [22]: %timeit in1d(a, b)  # Numpy-like implementation
1.71 ms ± 59.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
So maybe for small inputs, here some heuristics have to be used, the current approach should still be used?