@asi1024 2020-09-24
@mrkwjc Thank you for the PR!
The memory usage explosion in `ar1_broadcast == ar2_broadcast` seems to be avoidable by using `cupy.ReductionKernel`. I will compare the performance numbers between this PR and ReductionKernel version.