@asi1024 2020-09-28
I confirmed that this PR improves GPU performance, but causes host/device synchronization.
```
master    :    CPU:   31.325 us   +/-37.739 (min:   27.504 / max:  662.737) us     GPU-0:16218.380 us   +/-52.691 (min:16190.464 / max:16840.704) us
this PR   :    CPU: 1228.504 us   +/-82.605 (min: 1207.951 / max: 5493.522) us     GPU-0: 1366.806 us   +/-80.935 (min: 1344.512 / max: 5547.008) us
```